make frktest platform-agnostic (#8)
This change makes the reporter platform-agnostic, which was the only part of the library that was not. frktest should now work in any luau runtime, although text will be not be colored in non-lune runtimes.

It is however trivial to add coloring for other runtimes and I plan to add it for Zune and Seal

To make it a non-breaking change, I kept the old lune_console_reporter file and init object, but added a deprecated message

This also probably means I no longer need _pesde_init, and can just include the reporter in a new init that can be used by both pesde and wally.